Move visibility tests to a separate file
